+++ packages/clerk-js/src/ui/contexts/components/Plans.tsx
@@ -2,6 +2,7 @@ import {
   __experimental_usePaymentAttempts,
   __experimental_usePaymentMethods,
   __experimental_useStatements,
+  __experimental_useSubscriptionItems,
   useClerk,
   useOrganization,
   useSession,
@@ -16,7 +17,6 @@ import type {
 import { useCallback, useMemo } from 'react';
 import useSWR from 'swr';
 
-import { CommerceSubscription } from '@/core/resources/CommerceSubscription';
 import { getClosestProfileScrollBox } from '@/ui/utils/getClosestProfileScrollBox';
 
 import type { LocalizationKey } from '../../localization';
@@ -72,58 +72,14 @@ export const useStatements = (params?: { mode: 'cache' }) => {
 };
 
 export const useSubscriptions = () => {
-  const { billing } = useClerk();
-  const { organization } = useOrganization();
-  const { user, isSignedIn } = useUser();
   const subscriberType = useSubscriberTypeContext();
-  const { data: plans } = usePlans();
 
-  const { data: _subscriptions, ...rest } = useSWR(
-    {
-      key: `commerce-subscriptions`,
-      userId: user?.id,
-      args: { orgId: subscriberType === 'org' ? organization?.id : undefined },
-    },
-    ({ args, userId }) => (userId ? billing.getSubscriptions(args) : undefined),
-    dedupeOptions,
-  );
-
-  const subscriptions = useMemo(() => {
-    if (!_subscriptions) {
-      return [];
-    }
-    const defaultFreePlan = plans?.find(plan => plan.hasBaseFee === false && plan.amount === 0);
-
-    // are we signed in, is there a default free plan, and should it be shown as active or upcoming? then add an implicit subscription
-    if (
-      isSignedIn &&
-      defaultFreePlan &&
-      (_subscriptions.data.length === 0 || !_subscriptions.data.some(subscription => !subscription.canceledAtDate))
-    ) {
-      const canceledSubscription = _subscriptions.data.find(subscription => subscription.canceledAtDate);
-      return [
-        ..._subscriptions.data,
-        new CommerceSubscription({
-          object: 'commerce_subscription',
-          id: '__implicit_default_plan_subscription__',
-          payment_source_id: '',
-          plan: defaultFreePlan.__internal_toSnapshot(),
-          plan_period: 'month',
-          canceled_at: null,
-          status: _subscriptions.data.length === 0 ? 'active' : 'upcoming',
-          created_at: canceledSubscription?.periodEndDate?.getTime() || 0,
-          period_start: canceledSubscription?.periodEndDate?.getTime() || 0,
-          period_end: 0,
-        }),
-      ];
-    }
-    return _subscriptions.data;
-  }, [_subscriptions, plans, isSignedIn]);
-
-  return {
-    data: subscriptions,
-    ...rest,
-  };
+  return __experimental_useSubscriptionItems({
+    for: subscriberType === 'org' ? 'organization' : 'user',
+    initialPage: 1,
+    pageSize: 10,
+    keepPreviousData: true,
+  });
 };
 
 export const usePlans = () => {
@@ -143,7 +99,6 @@ export const usePlans = () => {
 type HandleSelectPlanProps = {
   plan: CommercePlanResource;
   planPeriod: CommerceSubscriptionPlanPeriod;
-  onSubscriptionChange?: () => void;
   mode?: 'modal' | 'mounted';
   event?: React.MouseEvent<HTMLElement>;
   appearance?: Appearance;
@@ -167,7 +122,7 @@ export const usePlansContext = () => {
     return false;
   }, [clerk, subscriberType]);
 
-  const { data: subscriptions, mutate: mutateSubscriptions } = useSubscriptions();
+  const { data: subscriptions, revalidate: revalidateSubscriptions } = useSubscriptions();
 
   // Invalidates cache but does not fetch immediately
   const { data: plans, mutate: mutatePlans } = useSWR<Awaited<ReturnType<typeof clerk.billing.getPlans>>>({
@@ -182,11 +137,11 @@ export const usePlansContext = () => {
 
   const revalidateAll = useCallback(() => {
     // Revalidate the plans and subscriptions
-    void mutateSubscriptions();
+    void revalidateSubscriptions();
     void mutatePlans();
     void revalidateStatements();
     void revalidatePaymentSources();
-  }, [mutateSubscriptions, mutatePlans, revalidateStatements, revalidatePaymentSources]);
+  }, [revalidateSubscriptions, mutatePlans, revalidateStatements, revalidatePaymentSources]);
 
   // should the default plan be shown as active
   const isDefaultPlanImplicitlyActiveOrUpcoming = useMemo(() => {
@@ -244,13 +199,6 @@ export const usePlansContext = () => {
     },
     [activeOrUpcomingSubscription],
   );
-
-  // should the default plan be shown as active
-  const upcomingSubscriptionsExist = useMemo(() => {
-    return (
-      subscriptions.some(subscription => subscription.status === 'upcoming') || isDefaultPlanImplicitlyActiveOrUpcoming
-    );
-  }, [subscriptions, isDefaultPlanImplicitlyActiveOrUpcoming]);
 
   // return the CTA button props for a plan
   const buttonPropsForPlan = useCallback(
@@ -332,6 +280,10 @@ export const usePlansContext = () => {
   );
 
   const captionForSubscription = useCallback((subscription: CommerceSubscriptionResource) => {
+    if (subscription.pastDueAt) {
+      return localizationKeys('badge__pastDueAt', { date: subscription.pastDueAt });
+    }
+
     if (subscription.status === 'upcoming') {
       return localizationKeys('badge__startsAt', { date: subscription.periodStartDate });
     }
@@ -345,55 +297,44 @@ export const usePlansContext = () => {
     return;
   }, []);
 
+  const openSubscriptionDetails = useCallback(
+    (event?: React.MouseEvent<HTMLElement>) => {
+      const portalRoot = getClosestProfileScrollBox('modal', event);
+      clerk.__internal_openSubscriptionDetails({
+        for: subscriberType,
+        onSubscriptionCancel: () => {
+          revalidateAll();
+        },
+        portalRoot,
+      });
+    },
+    [clerk, subscriberType, revalidateAll],
+  );
+
   // handle the selection of a plan, either by opening the subscription details or checkout
   const handleSelectPlan = useCallback(
-    ({
-      plan,
-      planPeriod,
-      onSubscriptionChange,
-      mode = 'mounted',
-      event,
-      appearance,
-      newSubscriptionRedirectUrl,
-    }: HandleSelectPlanProps) => {
-      const subscription = activeOrUpcomingSubscriptionWithPlanPeriod(plan, planPeriod);
-
+    ({ plan, planPeriod, mode = 'mounted', event, appearance, newSubscriptionRedirectUrl }: HandleSelectPlanProps) => {
       const portalRoot = getClosestProfileScrollBox(mode, event);
 
-      if (subscription && subscription.planPeriod === planPeriod && !subscription.canceledAtDate) {
-        clerk.__internal_openPlanDetails({
-          plan,
-          initialPlanPeriod: planPeriod,
-          subscriberType,
-          onSubscriptionCancel: () => {
-            revalidateAll();
-            onSubscriptionChange?.();
-          },
-          appearance,
-          portalRoot,
-        });
-      } else {
-        clerk.__internal_openCheckout({
-          planId: plan.id,
-          // if the plan doesn't support annual, use monthly
-          planPeriod: planPeriod === 'annual' && plan.annualMonthlyAmount === 0 ? 'month' : planPeriod,
-          subscriberType,
-          onSubscriptionComplete: () => {
-            revalidateAll();
-            onSubscriptionChange?.();
-          },
-          onClose: () => {
-            if (session?.id) {
-              void clerk.setActive({ session: session.id });
-            }
-          },
-          appearance,
-          portalRoot,
-          newSubscriptionRedirectUrl,
-        });
-      }
+      clerk.__internal_openCheckout({
+        planId: plan.id,
+        // if the plan doesn't support annual, use monthly
+        planPeriod: planPeriod === 'annual' && plan.annualMonthlyAmount === 0 ? 'month' : planPeriod,
+        subscriberType,
+        onSubscriptionComplete: () => {
+          revalidateAll();
+        },
+        onClose: () => {
+          if (session?.id) {
+            void clerk.setActive({ session: session.id });
+          }
+        },
+        appearance,
+        portalRoot,
+        newSubscriptionRedirectUrl,
+      });
     },
-    [clerk, revalidateAll, activeOrUpcomingSubscription, subscriberType, session?.id],
+    [clerk, revalidateAll, subscriberType, session?.id],
   );
 
   const defaultFreePlan = useMemo(() => {
@@ -406,10 +347,10 @@ export const usePlansContext = () => {
     activeOrUpcomingSubscriptionBasedOnPlanPeriod: activeOrUpcomingSubscriptionWithPlanPeriod,
     isDefaultPlanImplicitlyActiveOrUpcoming,
     handleSelectPlan,
+    openSubscriptionDetails,
     buttonPropsForPlan,
     canManageSubscription,
     captionForSubscription,
-    upcomingSubscriptionsExist,
     defaultFreePlan,
     revalidateAll,
   };