+++ packages/shared/src/react/commerce.tsx
@@ -7,6 +7,7 @@ import useSWR from 'swr';
 import useSWRMutation from 'swr/mutation';
 
 import { createContextAndHook } from './hooks/createContextAndHook';
+import type { useCheckout } from './hooks/useCheckout';
 import { useClerk } from './hooks/useClerk';
 import { useOrganization } from './hooks/useOrganization';
 import { useUser } from './hooks/useUser';
@@ -135,7 +136,7 @@ type internalStripeAppearance = {
 };
 
 type PaymentElementProviderProps = {
-  checkout?: CommerceCheckoutResource;
+  checkout?: CommerceCheckoutResource | ReturnType<typeof useCheckout>['checkout'];
   stripeAppearance?: internalStripeAppearance;
   // TODO(@COMMERCE): What can we do to remove this ?
   for: 'org' | 'user';
@@ -173,12 +174,10 @@ const PropsProvider = ({ children, ...props }: PropsWithChildren<PaymentElementP
     <PaymentElementContext.Provider
       value={{
         value: {
+          ...props,
           ...utils,
           setIsPaymentElementReady,
           isPaymentElementReady,
-          checkout: props.checkout,
-          paymentDescription: props.paymentDescription,
-          for: props.for,
         },
       }}
     >
@@ -234,7 +233,7 @@ const PaymentElement = ({ fallback }: { fallback?: ReactNode }) => {
   const environment = useInternalEnvironment();
 
   const applePay = useMemo(() => {
-    if (!checkout) {
+    if (!checkout || !checkout.totals || !checkout.plan) {
       return undefined;
     }