+++ packages/expo/CHANGELOG.md
@@ -1,5 +1,55 @@
 # Change Log
 
+## 2.14.7
+
+### Patch Changes
+
+- Updated dependencies [[`2a90b68`](https://github.com/clerk/javascript/commit/2a90b689550ae960496c9292ca23e0225e3425cd), [`af50905`](https://github.com/clerk/javascript/commit/af50905ea497ed3286c8c4c374498e06ca6ee82b)]:
+  - @clerk/clerk-js@5.73.2
+  - @clerk/clerk-react@5.35.2
+  - @clerk/types@4.67.0
+  - @clerk/shared@3.12.3
+
+## 2.14.6
+
+### Patch Changes
+
+- Updated dependencies [[`8ee859c`](https://github.com/clerk/javascript/commit/8ee859ce00d1d5747c14a80fe7166303e64a4f1f), [`c666c80`](https://github.com/clerk/javascript/commit/c666c80be32ee76066f24b34b520fb6ba1ff6d43)]:
+  - @clerk/shared@3.12.2
+  - @clerk/types@4.66.1
+  - @clerk/clerk-js@5.73.1
+  - @clerk/clerk-react@5.35.1
+
+## 2.14.5
+
+### Patch Changes
+
+- Updated dependencies [[`025e304`](https://github.com/clerk/javascript/commit/025e304c4d6402dfd750ee51ac9c8fc2dea1f353), [`dedf487`](https://github.com/clerk/javascript/commit/dedf48703986d547d5b28155b0182a51030cffeb), [`b96114e`](https://github.com/clerk/javascript/commit/b96114e438638896ba536bb7a17b09cdadcd9407), [`fdd4f0b`](https://github.com/clerk/javascript/commit/fdd4f0b981ea47d9d4fa597f38e49ed2a057cfc7)]:
+  - @clerk/clerk-js@5.73.0
+  - @clerk/types@4.66.0
+  - @clerk/clerk-react@5.35.0
+  - @clerk/shared@3.12.1
+
+## 2.14.4
+
+### Patch Changes
+
+- Updated dependencies [[`2be6a53`](https://github.com/clerk/javascript/commit/2be6a53959cb8a3127c2eb5d1aeb4248872d2c24), [`f6a1c35`](https://github.com/clerk/javascript/commit/f6a1c35bd5fb4bd2a3cd45bdaf9defe6be59d4a9), [`d97dd35`](https://github.com/clerk/javascript/commit/d97dd3598ae84d6985eeee53c4812204dea54b6b), [`6826d0b`](https://github.com/clerk/javascript/commit/6826d0bbd03e844d49224565878a4326684f06b4), [`f6a1c35`](https://github.com/clerk/javascript/commit/f6a1c35bd5fb4bd2a3cd45bdaf9defe6be59d4a9), [`97a07f7`](https://github.com/clerk/javascript/commit/97a07f78b4b0c3dc701a2610097ec7d6232f79e7), [`32a4952`](https://github.com/clerk/javascript/commit/32a4952a610d74f64da7837fc174eafd4855c671), [`ab31ae8`](https://github.com/clerk/javascript/commit/ab31ae84a3e6ec45bfe38bc89baefabb0bec81b6)]:
+  - @clerk/types@4.65.0
+  - @clerk/shared@3.12.0
+  - @clerk/clerk-react@5.34.0
+  - @clerk/clerk-js@5.72.0
+
+## 2.14.3
+
+### Patch Changes
+
+- Updated dependencies [[`50fc006`](https://github.com/clerk/javascript/commit/50fc006180519ffbdb6dfb5b9c6925482fb33849), [`f42c4fe`](https://github.com/clerk/javascript/commit/f42c4fedfdab873129b876eba38b3677f190b460), [`22be9f0`](https://github.com/clerk/javascript/commit/22be9f06d3419baa32212e2e5a5fa91d40240f7d), [`4ab82f7`](https://github.com/clerk/javascript/commit/4ab82f7e112dc6e904a33ebf21d4d3ef4bb07319), [`ec207dc`](https://github.com/clerk/javascript/commit/ec207dcd2a13340cfa4e3b80d3d52d1b4e7d5f23), [`ec207dc`](https://github.com/clerk/javascript/commit/ec207dcd2a13340cfa4e3b80d3d52d1b4e7d5f23), [`7f341dc`](https://github.com/clerk/javascript/commit/7f341dc7ed4982324e6dae59e803f42e794f2e2f), [`ec207dc`](https://github.com/clerk/javascript/commit/ec207dcd2a13340cfa4e3b80d3d52d1b4e7d5f23), [`6b4b27e`](https://github.com/clerk/javascript/commit/6b4b27e9d78417218280ef08b376883711eff47e), [`2fbdad0`](https://github.com/clerk/javascript/commit/2fbdad0bd93618bcf3e9a1aa94383eb2785c157c), [`0e0cc1f`](https://github.com/clerk/javascript/commit/0e0cc1fa85347d727a4fd3718fe45b0f0244ddd9), [`d34205c`](https://github.com/clerk/javascript/commit/d34205c7b85329f04948e2236b4a7fefeaf62884)]:
+  - @clerk/clerk-js@5.71.0
+  - @clerk/types@4.64.0
+  - @clerk/clerk-react@5.33.0
+  - @clerk/shared@3.11.0
+
 ## 2.14.2
 
 ### Patch Changes